# backend/openapi.yaml
openapi: 3.1.0
info:
  title: FinBot API
  version: 0.1.0
paths:
  /auth/register:
    post:
      tags:
      - auth
      summary: Register
      description: 'Create a new user account (role: user). Admins are seeded from
        env vars.'
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
      - auth
      summary: Login
      description: Authenticate with email and password. Returns a JWT bearer token.
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/me:
    get:
      tags:
      - auth
      summary: Me
      description: Return the authenticated user's profile.
      operationId: me_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /chat:
    post:
      tags:
      - chat
      summary: Chat
      description: 'Main chat endpoint. Checks semantic cache first, then RAG context,

        then calls the LLM agent with tool calling support.

        Returns: {reply, tool_used, from_cache}'
      operationId: chat_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /transcribe:
    post:
      tags:
      - chat
      summary: Transcribe
      description: Speech-to-Text via Whisper. Accepts .mp3, .wav, .ogg, .webm (up
        to 25 MB).
      operationId: transcribe_transcribe_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_transcribe_transcribe_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /speak:
    post:
      tags:
      - chat
      summary: Speak
      description: Text-to-Speech via OpenAI TTS. Returns audio/mpeg binary stream.
      operationId: speak_speak_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeakRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /reset:
    post:
      tags:
      - chat
      summary: Reset
      description: Clear session memory.
      operationId: reset_reset_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /health:
    get:
      tags:
      - system
      summary: Health
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /admin/analytics:
    get:
      tags:
      - admin
      summary: Admin Analytics
      description: Tool call counts, cache hit rate and top queries.
      operationId: admin_analytics_admin_analytics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /admin/config:
    post:
      tags:
      - admin
      summary: Admin Config
      description: Write env vars to .env and update os.environ. SIMILARITY_THRESHOLD
        applies immediately.
      operationId: admin_config_admin_config_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminConfigRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /admin/rag/add:
    post:
      tags:
      - admin
      summary: Admin Rag Add
      description: Index a new URL and merge it into the in-memory FAISS index.
      operationId: admin_rag_add_admin_rag_add_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagAddRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /admin/rag/reindex:
    post:
      tags:
      - admin
      summary: Admin Rag Reindex
      description: Reindex one URL, or reset and re-ingest all tracked URLs.
      operationId: admin_rag_reindex_admin_rag_reindex_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagReindexRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /admin/rag:
    delete:
      tags:
      - admin
      summary: Admin Rag Delete
      description: Remove a URL from the index (reset + re-ingest remaining URLs).
      operationId: admin_rag_delete_admin_rag_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagDeleteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /admin/cache:
    delete:
      tags:
      - admin
      summary: Admin Cache Clear
      description: Truncate the semantic cache (memory + DB).
      operationId: admin_cache_clear_admin_cache_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
components:
  schemas:
    AdminConfigRequest:
      properties:
        env:
          additionalProperties:
            type: string
          type: object
          title: Env
      type: object
      required:
      - env
      title: AdminConfigRequest
    Body_transcribe_transcribe_post:
      properties:
        audio:
          type: string
          contentMediaType: application/octet-stream
          title: Audio
      type: object
      required:
      - audio
      title: Body_transcribe_transcribe_post
    ChatRequest:
      properties:
        message:
          type: string
          title: Message
        image_b64:
          anyOf:
          - type: string
          - type: 'null'
          title: Image B64
      type: object
      required:
      - message
      title: ChatRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    RagAddRequest:
      properties:
        url:
          type: string
          title: Url
      type: object
      required:
      - url
      title: RagAddRequest
    RagDeleteRequest:
      properties:
        url:
          type: string
          title: Url
      type: object
      required:
      - url
      title: RagDeleteRequest
    RagReindexRequest:
      properties:
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
      type: object
      title: RagReindexRequest
    RegisterRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: RegisterRequest
    SpeakRequest:
      properties:
        text:
          type: string
          title: Text
      type: object
      required:
      - text
      title: SpeakRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer